+++ packages/js/src/dashboard/contexts/alerts-context.js
@@ -1,0 +1,6 @@
+import { createContext } from "@wordpress/element";
+
+/**
+ * The context for the alerts.
+ */
+export const AlertsContext = createContext( { Icon: null, bulletClass: "", iconClass: "" } );
